+++ src/core/ports/index.ts
@@ -1,5 +1,1 @@
-export * from './jsonplaceholder.integration';
-export * from './todo.cache';
-export * from './todo.producer';
-export * from './todo.collection';
-export * from './todo.repository';
+export * from './todo.service';
