+++ src/core/ports/todo.service.ts
@@ -1,5 +1,5 @@
-import { JsonPlaceHolderUser } from '../../types';
-import { Todo } from '../../entities';
+import { Todo } from '../entities';
+import { JsonPlaceHolderUser } from '../types';
 
 export interface ITodoService {
   create(dto: Todo): Promise<string>;
