--- src/core/services/index.ts
@@ -1,2 +1,0 @@
-export * from './todo/todo.interface';
-export * from './todo/todo';
