--- src/infra/integrations/amqp/index.ts
@@ -1,6 +1,0 @@
-export * from './type';
-export * from './interface';
-export * from './client';
-export * from './producer/producer';
-export * from './producer/types';
-export * from './producer/todo/todo';
