+++ src/infra/integrations/amqp/interface.ts
@@ -1,4 +1,4 @@
-import { AMQPPublishData } from '.';
+import { AMQPPublishData } from '..';
 
 export interface IAmqp {
   publish(data: AMQPPublishData): Promise<boolean>;
