+++ src/infra/integrations/amqp/producer/producer.ts
@@ -1,4 +1,4 @@
-import { ProducerConfig } from '..';
+import { ProducerConfig } from '../..';
 
 export abstract class Producer {
   abstract producer_config: ProducerConfig;
