--- src/infra/integrations/http/index.ts
@@ -1,5 +1,0 @@
-export * from './client';
-export * from './interfaces';
-export * from './enum';
-export * from './jsonplaceholder/jsonplaceholder';
-export * from './jsonplaceholder/jsonplaceholder.response';
