+++ src/infra/integrations/http/jsonplaceholder/jsonplaceholder.interface.ts
@@ -1,4 +1,4 @@
-import { JsonPlaceHolderUser } from '../types';
+import { JsonPlaceHolderUser } from '../../../../core/types';
 
 export interface IJsonPlaceHolderIntegration {
   getUser(id: string): Promise<JsonPlaceHolderUser>;
