+++ src/infra/repositories/index.ts
@@ -1,8 +1,11 @@
 export * from './relational/repository';
 export * from './relational/types';
 export * from './relational/todo/todo';
+export * from './relational/todo/todo.interface';
 export * from './cache/client';
 export * from './cache/interface';
 export * from './cache/todo/todo';
+export * from './cache/todo/todo.interface';
 export * from './no_sql/todos/schema';
 export * from './no_sql/todos/todo';
+export * from './no_sql/todos/todo.interface';
