--- src/infra/repositories/no_sql/index.ts
@@ -1,2 +1,0 @@
-export * from './todos/todo';
-export * from './todos/schema';
