+++ src/infra/repositories/no_sql/todos/todo.interface.ts
@@ -1,4 +1,4 @@
-import { Todo } from '../entities';
+import { Todo } from '../../../../core/entities';
 
 export interface ITodoCollection {
   save(todo: Todo): Promise<void>;
