+++ src/infra/repositories/relational/todo/todo.ts
@@ -1,7 +1,6 @@
 import Knex from 'knex';
-import { Repository, TableConfig } from '../..';
+import { Repository, TableConfig, ITodoRepository } from '../..';
 import { Todo } from '../../../../core/entities';
-import { ITodoRepository } from '../../../../core/ports';
 
 export class TodoRepository
   extends Repository<Todo>
